improved testing so that pdflatex is used to validate the results made all splint suggestions to eqn.c
git-svn-id: svn://svn.code.sf.net/p/rtf2latex2e/code@76 ec8729b6-1b83-4082-b6c8-10560a4245e1

+++ src/eqn.h
@@ -13,7 +13,7 @@ typedef struct {
 
 typedef struct {
     struct MT_OBJLIST *next;
-    int tag;
+    uint8_t tag;
     void *obj_ptr;
 } MT_OBJLIST;
